Refactor recipe building logic in Heroic Green Steel pages
- Add `useRecipeBuilder` hook for reusable recipe handling logic.
- Remove duplicate recipe handling logic from `IngredientList.tsx`.
- Replace manual state updates with `useRecipeBuilder` in `IngredientList.tsx`.
